requirements.txt
slackClient
python-binance